refactor: colocate settings panel rendering in its citizen file
